+++ src/frameworks/next/index.ts
@@ -28,6 +28,7 @@ import {
   findDependency,
   validateLocales,
   getNodeModuleBin,
+  getBuildScript,
 } from "../utils";
 import {
   BuildResult,
@@ -87,7 +88,7 @@ import { getAllSiteDomains, getDeploymentDomain } from "../../hosting/api";
 import { logger } from "../../logger";
 import { parseStrict } from "../../functions/env";
 
-const DEFAULT_BUILD_SCRIPT = ["next build"];
+const DEFAULT_BUILD_SCRIPT = ["next build", "next build --webpack"];
 const PUBLIC_DIR = "public";
 
 export const supportedRange = "12 - 15.0";
@@ -119,10 +120,13 @@ export async function discover(dir: string) {
  */
 export async function build(
   dir: string,
-  target: string,
+  _target: string,
   context?: FrameworkContext,
 ): Promise<BuildResult> {
-  await warnIfCustomBuildScript(dir, name, DEFAULT_BUILD_SCRIPT);
+  const buildScript = await getBuildScript(join(dir, "package.json"));
+  if (buildScript) {
+    warnIfCustomBuildScript(buildScript, name, DEFAULT_BUILD_SCRIPT);
+  }
 
   const reactVersion = getReactVersion(dir);
   if (reactVersion && gte(reactVersion, "18.0.0")) {
@@ -162,7 +166,11 @@ export async function build(
   const cli = getNodeModuleBin("next", dir);
 
   const nextBuild = new Promise((resolve, reject) => {
-    const buildProcess = spawn(cli, ["build"], { cwd: dir, env });
+    const buildProcess = spawn(
+      cli,
+      ["build", ...(buildScript?.includes("--webpack") ? ["--webpack"] : [])],
+      { cwd: dir, env },
+    );
     buildProcess.stdout?.on("data", (data) => logger.info(data.toString()));
     buildProcess.stderr?.on("data", (data) => logger.info(data.toString()));
     buildProcess.on("error", (err) => {