+++ src/frameworks/utils.ts
@@ -51,19 +51,31 @@ export function readJSON<JsonType = any>(
  * Prints a warning if the build script in package.json
  * contains anything other than allowedBuildScripts.
  */
-export async function warnIfCustomBuildScript(
-  dir: string,
+export function warnIfCustomBuildScript(
+  buildScript: string,
   framework: string,
   defaultBuildScripts: string[],
-): Promise<void> {
-  const packageJsonBuffer = await readFile(join(dir, "package.json"));
-  const packageJson = JSON.parse(packageJsonBuffer.toString());
-  const buildScript = packageJson.scripts?.build;
-
-  if (buildScript && !defaultBuildScripts.includes(buildScript)) {
+): void {
+  if (!defaultBuildScripts.includes(buildScript)) {
     console.warn(
       `\nWARNING: Your package.json contains a custom build that is being ignored. Only the ${framework} default build script (e.g, "${defaultBuildScripts[0]}") is respected. If you have a more advanced build process you should build a custom integration https://firebase.google.com/docs/hosting/express\n`,
     );
+  }
+}
+
+/**
+ * Get the build script from package.json.
+ *
+ * @param packageJsonPath - The path to the package.json file.
+ * @returns The build script from package.json, or undefined if the build script or package.json is not found.
+ */
+export async function getBuildScript(packageJsonPath: string): Promise<string | undefined> {
+  try {
+    const packageJsonBuffer = await readFile(packageJsonPath);
+    const packageJson = JSON.parse(packageJsonBuffer.toString());
+    return packageJson.scripts?.build;
+  } catch (error) {
+    return undefined;
   }
 }