+++ CHANGELOG.md
@@ -1,2 +1,3 @@
+- Add support for the --webpack flag in Next.js build (#9761)
 - Fix the bug when Data Connect emulator hangs with PGlite. (Issue #9756) #9771
 - Updated to v3.1.0 of the Data Connect emulator, which adds support for using `distinct: true` in combination with `orderBy`, and other internal improvements (#9778).